requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
pint>=0.21.0
pydantic>=2.0.0
python-dotenv>=1.0.0
groq>=0.32.0
